refactor: extract artist profile form section from profile page
